Rename the module and fix the model training
