+++ .gitignore
@@ -2,6 +2,7 @@
 datasets/
 checkpoints/
 models/
+results/
 
 ### Python template
 # Byte-compiled / optimized / DLL files
