+++ aocr/model/cnn.py
@@ -1,11 +1,4 @@
-__author__ = 'moonkey'
-
-#from keras import models, layers
-import logging
 import numpy as np
-# from src.data_util.synth_prepare import SynthGen
-
-#import keras.backend as K
 import tensorflow as tf
 
 
@@ -24,6 +17,7 @@ def var_random(name, shape, regularizable=False):
             tf.add_to_collection(tf.GraphKeys.REGULARIZATION_LOSSES, tf.nn.l2_loss(v))
     return v
 
+
 def max_2x2pool(incoming, name):
     '''
     max pooling on 2 dims.
@@ -34,6 +28,7 @@ def max_2x2pool(incoming, name):
     with tf.variable_scope(name):
         return tf.nn.max_pool(incoming, ksize=(1, 2, 2, 1), strides=(1, 2, 2, 1), padding='VALID')
 
+
 def max_2x1pool(incoming, name):
     '''
     max pooling only on image width
@@ -43,6 +38,7 @@ def max_2x1pool(incoming, name):
     '''
     with tf.variable_scope(name):
         return tf.nn.max_pool(incoming, ksize=(1, 2, 1, 1), strides=(1, 2, 1, 1), padding='VALID')
+
 
 def ConvRelu(incoming, num_filters, filter_size, name):
     '''
@@ -87,14 +83,16 @@ def ConvReluBN(incoming, num_filters, filter_size, name, is_training, padding_ty
     with tf.variable_scope(name):
         conv_W = var_random('W', tuple(filter_size) + (num_filters_from, num_filters), regularizable=True)
 
-        after_conv = tf.nn.conv2d(incoming, conv_W, strides=(1,1,1,1), padding=padding_type)
+        after_conv = tf.nn.conv2d(incoming, conv_W, strides=(1, 1, 1, 1), padding=padding_type)
 
         after_bn = batch_norm(after_conv, is_training)
 
         return tf.nn.relu(after_bn)
 
+
 def dropout(incoming, is_training, keep_prob=0.5):
     return tf.contrib.layers.dropout(incoming, keep_prob=keep_prob, is_training=is_training)
+
 
 def tf_create_attention_map(incoming):
     '''
@@ -106,6 +104,7 @@ def tf_create_attention_map(incoming):
     print("shape of incoming is: {}".format(incoming.get_shape()))
     print(shape)
     return tf.reshape(incoming, (-1, np.prod(shape[1:3]), shape[3]))
+
 
 class CNN(object):
     """
@@ -146,7 +145,7 @@ class CNN(object):
 
         print('CNN outdim before squeeze: {}'.format(net.get_shape()))  # 1x32x100 -> 24x512
 
-        net = tf.squeeze(net,axis=1)
+        net = tf.squeeze(net, axis=1)
 
         print('CNN outdim: {}'.format(net.get_shape()))
         self.model = net
@@ -160,5 +159,3 @@ class CNN(object):
     '''
     def save(self):
         pass
-
-
