+++ aocr/util/data_gen.py
@@ -1,5 +1,3 @@
-__author__ = 'moonkey, emedvedev'
-
 import os
 import math
 
@@ -13,20 +11,18 @@ from .bucketdata import BucketData
 
 
 class DataGen(object):
-    _GO = 1
-    _EOS = 2
-
-    IMG_HEIGHT = 32
+    GO_ID = 1
+    EOS_ID = 2
+    IMAGE_HEIGHT = 32
 
     def __init__(self,
-                 data_root, annotation_fn,
+                 annotation_fn,
                  evaluate=False,
                  valid_target_len=float('inf'),
                  img_width_range=(12, 320),
                  word_len=30,
                  epochs=1000):
         """
-        :param data_root:
         :param annotation_fn:
         :param lexicon_fn:
         :param valid_target_len:
@@ -35,16 +31,14 @@ class DataGen(object):
         :param epochs:
         :return:
         """
-        self.data_root = data_root
         self.epochs = epochs
-        self.image_height = self.IMG_HEIGHT
         self.valid_target_len = valid_target_len
         self.bucket_min_width, self.bucket_max_width = img_width_range
 
         if os.path.exists(annotation_fn):
             self.annotation_path = annotation_fn
         else:
-            self.annotation_path = os.path.join(data_root, annotation_fn)
+            raise IOError("The .tfrecords file %s does not exist." % annotation_fn)
 
         if evaluate:
             self.bucket_specs = [(int(math.floor(64 / 4)), int(word_len + 2)),
@@ -88,15 +82,16 @@ class DataGen(object):
                 while not coord.should_stop():
                     raw_images, raw_labels = sess.run([images, labels])
                     for img, lex in zip(raw_images, raw_labels):
-                        _, word = self.read_data(None, lex)
+                        word = self.convert_lex(lex)
                         if valid_target_len < float('inf'):
                             word = word[:valid_target_len + 1]
 
                         img_data = Image.open(StringIO(img))
-                        width, _ = img_data.size
+                        width, height = img_data.size
+                        resized_width = math.floor(float(width) / height * self.IMAGE_HEIGHT)
 
-                        b_idx = min(width, self.bucket_max_width)
-                        bucket_size = self.bucket_data[b_idx].append(img, width, word, lex)
+                        b_idx = min(resized_width, self.bucket_max_width)
+                        bucket_size = self.bucket_data[b_idx].append(img, resized_width, word, lex)
                         if bucket_size >= batch_size:
                             bucket = self.bucket_data[b_idx].flush_out(
                                 self.bucket_specs,
@@ -105,7 +100,7 @@ class DataGen(object):
                             if bucket is not None:
                                 yield bucket
                             else:
-                                assert False, 'no valid bucket of width %d' % width
+                                assert False, 'no valid bucket of width %d' % resized_width
 
             finally:
                 coord.request_stop()
@@ -113,18 +108,18 @@ class DataGen(object):
 
         self.clear()
 
-    def read_data(self, img, lex):
+    def convert_lex(self, lex):
         assert lex and len(lex) < self.bucket_specs[-1][1]
 
-        word = [self._GO]
+        word = [self.GO_ID]
         for char in lex:
             assert 96 < ord(char) < 123 or 47 < ord(char) < 58
             word.append(
                 ord(char) - 97 + 13 if ord(char) > 96 else ord(char) - 48 + 3)
-        word.append(self._EOS)
+        word.append(self.EOS_ID)
         word = np.array(word, dtype=np.int32)
 
-        return img, word
+        return word
 
 
 def parse_tfrecords(filename_queue):
@@ -134,6 +129,6 @@ def parse_tfrecords(filename_queue):
         serialized_example,
         features={
             'image': tf.FixedLenFeature([], tf.string),
-            'label': tf.FixedLenFeature([], tf.string),
+            'answer': tf.FixedLenFeature([], tf.string),
         })
-    return features['image'], features['label']
+    return features['image'], features['answer']
