--- src/data_util/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'moonkey'
